Moved editor into separate component
